refactor: move out Capture type to a dedicated file
